+++ R/mdirt.R
@@ -27,6 +27,17 @@
 #' grid of \eqn{\theta} values will allow the estimation of similar models (multidimensional
 #' discrete models, grade of membership, etc.). See the examples below.
 #'
+#' When the \code{item.Q} for is utilized, the above equation can be understood as
+#'
+#' \deqn{P(x = k|\theta_1, \theta_2, a1, a2) = \frac{exp(a1 \theta_1 Q_{j1} + a2 \theta_2 Q_{j2})}{
+#'   \sum_j^K exp(a1 \theta_1 Q_{j1} + a2 \theta_2 Q_{j2})}}
+#'
+#' where by construction \code{Q} is a \eqn{K_i \times A} matrix indicating whether the category should
+#' be modeled according to the latent class structure. For the standard latent class model, the Q-matrix
+#' has as many rows as categories, as many columns as the number of classes/attributes modeled,
+#' and consist of 0's in the first row and 1's elsewhere. This of course can be over-written by passing
+#' an alternative \code{item.Q} definition for each respective item.
+#'
 #' @param data a \code{matrix} or \code{data.frame} that consists of
 #'   numerically ordered data, with missing data coded as \code{NA}
 #' @param model number of mutually exclusive classes to fit, or alternatively a more specific
@@ -41,6 +52,13 @@
 #' @param itemtype a vector indicating the itemtype associated with each item.
 #'   For discrete models this is limited to only 'lca' or items defined using a
 #'   \code{\link{createItem}} definition
+#' @param item.Q a list of item-level Q-matricies indicating how the respective categories should be
+#'   modeled by the underlying attributes. Each matrix must represent a \eqn{K_i \times A} matrix,
+#'   where \eqn{K_i} represents the number of categories for the ith item, and \eqn{A} is the number
+#'   of attributes included in the \code{Theta} matrix; otherwise, a value of\code{NULL} will default
+#'   to a matrix consisting of 1's for each \eqn{K_i \times A} element except for the first row, which
+#'   contains only 0's for proper identification. Incidentally, the first row of each matrix \code{must}
+#'   contain only 0's so that the first category represents the reference category for identification
 #' @param GenRandomPars logical; use random starting values
 #' @param customTheta input passed to \code{technical = list(customTheta = ...)}, but is included
 #'   directly in this function for convenience. This input is most interesting for discrete latent models
@@ -268,8 +286,8 @@
 #'
 #'
 #' }
-mdirt <- function(data, model, customTheta = NULL, structure = NULL, nruns = 1, method = 'EM',
-                  covdata = NULL, formula = NULL, itemtype = 'lca',
+mdirt <- function(data, model, customTheta = NULL, structure = NULL, item.Q = NULL,
+                  nruns = 1, method = 'EM', covdata = NULL, formula = NULL, itemtype = 'lca',
                   optimizer = 'nlminb', return_max = TRUE, group = NULL, GenRandomPars = FALSE,
                   verbose = TRUE, pars = NULL, technical = list(), ...)
 {
@@ -292,7 +310,7 @@ mdirt <- function(data, model, customTheta = NULL, structure = NULL, nruns = 1, 
     mods <- myLapply(1:nruns, function(x, ...) return(ESTIMATION(...)), method=method,
                      latent.regression=latent.regression, structure=structure,
                      data=data, model=model, group=group, itemtype=itemtype, optimizer=optimizer,
-                     technical=technical, calcNull=FALSE, GenRandomPars=GenRandomPars,
+                     technical=technical, calcNull=FALSE, GenRandomPars=GenRandomPars, item.Q=item.Q,
                      dentype = 'discrete', verbose=ifelse(nruns > 1L, FALSE, verbose), pars=pars, ...)
     if(is(mods[[1L]], 'DiscreteClass')){
         for(i in 1:length(mods)) mods[[i]]@Call <- Call